refactor: 改为本地读取敏感词
