--- util/sensitive_filter.go
@@ -1,15 +1,0 @@
-package util
-
-import (
-	"github.com/importcjj/sensitive"
-)
-
-const NetWordDicUrl = "https://raw.githubusercontent.com/importcjj/sensitive/master/dict/dict.txt"
-
-var Filter *sensitive.Filter
-
-func InitFilter() error {
-	Filter = sensitive.New()
-	err := Filter.LoadNetWordDict(NetWordDicUrl)
-	return err
-}
